// client/src/common/types/hr/payroll/Payroll.ts
export type PayrollStatus =
  | "BROUILLON"
  | "VERIFICATION"
  | "APPROUVÉ"
  | "PAYÉ"
  | "ANNULÉ";
// ...
export type PayrollResultStatus = Exclude<PayrollStatus, "VERIFICATION"> | "VERIFIÉ";

export interface PayrollResult {
  companyId: string;
  _id?: string;
  generatedBy: string;
  payrollRunId?: string;
  month: number;
  year: number;
  employeeId?: string;
  firstName?: string;
  lastName?: string;
  department?: string;
  accountNumber?: string;
  baseSalary: number;
  grossSalary: number;
  taxableSalary: number;
  earnings: PayrollItem[];
  deductions: PayrollItem[];
  totalEarnings: number;
  totalDeductions: number;
  status: PayrollResultStatus;
  serverVersion?: number;
  cancelledAt?: string;
  cancelledBy?: string;
  cancelledByName?: string;
  verifiedAt?: string;
  verifiedBy?: string;
  verifiedByName?: string;
  approvedBy?: string;
  approvedByName?: string;
  approvedAt?: string;
  paidBy?: string;
  paidByName?: string;
  paidAt?: string;
  netSalary: number;
  createdAt?: string;
  updatedAt?: string;
  isDeleted?: number;
}

export interface PayrollItem {
  companyId: string;
  _id?: string;
  employeeId?: string;
  payrollResultId?: string;
  componentId?: string;
  name: string;
  displayName?: string;
  type: PayrollComponentType;
  amount: number;
  calculationMethod?: string;
  rate?: number;
  taxable?: number;
  quantity?: number;
  notes?: string;
  synced?: number;
  serverVersion?: number;
  createdAt?: string;
  updatedAt?: string;
  lastSyncedAt?: string;
  isDeleted?: number;
}

export type PayrollComponentType = "EARNING" | "DEDUCTION";